Revert "Optimize i18n language file directory structure (#1023)"
This reverts commit dc73f09c55ad54d6e6d8c4dab26373a6afa4d454.
