+++ backend/common/i18n.py
@@ -1,24 +1,14 @@
-import glob
-import json
-
-from pathlib import Path
 from typing import Any
-
-import yaml
 
 from starlette_context.errors import ContextDoesNotExistError
 
 from backend.common.context import ctx
 from backend.core.conf import settings
-from backend.core.path_conf import LOCALE_DIR
+from backend.locales.loader import locale_loader
 
 
 class I18n:
     """国际化管理器"""
-
-    def __init__(self) -> None:
-        self.locales: dict[str, dict[str, Any]] = {}
-        self.load_locales()
 
     @property
     def current_language(self) -> str:
@@ -33,29 +23,6 @@ class I18n:
         """设置当前请求的语言"""
         ctx.language = language
 
-    def load_locales(self) -> None:
-        """加载语言文本"""
-        patterns = [
-            LOCALE_DIR / '*.json',
-            LOCALE_DIR / '*.yaml',
-            LOCALE_DIR / '*.yml',
-        ]
-
-        lang_files = []
-
-        for pattern in patterns:
-            lang_files.extend(glob.glob(str(pattern)))
-
-        for lang_file in lang_files:
-            with open(lang_file, encoding='utf-8') as f:
-                lang = Path(lang_file).stem
-                file_type = Path(lang_file).suffix[1:]
-                match file_type:
-                    case 'json':
-                        self.locales[lang] = json.loads(f.read())
-                    case 'yaml' | 'yml':
-                        self.locales[lang] = yaml.full_load(f.read())
-
     def t(self, key: str, default: Any | None = None, **kwargs) -> str:
         """
         翻译函数
@@ -68,10 +35,10 @@ class I18n:
         keys = key.split('.')
 
         try:
-            translation = self.locales[self.current_language]
+            translation = locale_loader.locales[self.current_language]
         except KeyError:
             keys = 'error.language_not_found'.split('.')
-            translation = self.locales[settings.I18N_DEFAULT_LANGUAGE]
+            translation = locale_loader.locales[settings.I18N_DEFAULT_LANGUAGE]
 
         for k in keys:
             if isinstance(translation, dict) and k in list(translation.keys()):
